pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ulvek"
version = "0.1.0"
description = "Agentic Coding Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.9.0",
    "ag-ui-protocol>=0.1.8",
    "aiofiles>=24.1.0",
    "chonkie[add,code]>=1.1.1",
    "click>=8.2.1",
    "cryptography>=45.0.5",
    "fastapi[standard]>=0.116.1",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "langgraph>=0.5.4",
    "magika>=0.6.2",
    "mem0ai>=0.1.115",
    "pathspec>=0.12.1",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.4.7",
    "rich>=14.1.0",
    "rich-click>=1.8.9",
    "ripgrepy>=2.2.0",
    "sentence-transformers>=5.0.0",
    "sseclient>=0.0.27",
    "textual>=5.3.0",
    "typer>=0.16.0",
]

[project.scripts]
ulvek = "src.cli.main:main"
